* [TOB] DEV-3786: ID-7 (#13)

* updated sample collaterals and added crl timestamp check

* missing x509 expiration check

* [TOB] DEV-3793: JSON Collateral Content Hash (#14)

* minor changes on type declaration to stay consistent with Solidity types

* fmspc tcb content hash

* enclave identity content hash

* tee type in verified output is encoded little-endian, but we are keeping quote_version as big endian for on-chain compatibility

+++ src/types/enclave_identity.rs
@@ -142,10 +142,10 @@ pub struct EnclaveIdentityV2 {
 #[serde(rename_all = "camelCase")]
 pub struct EnclaveIdentityV2Inner {
     pub id: String,
-    pub version: u64,
+    pub version: u32,
     pub issue_date: String,
     pub next_update: String,
-    pub tcb_evaluation_data_number: u64,
+    pub tcb_evaluation_data_number: u32,
     pub miscselect: String,
     pub miscselect_mask: String,
     pub attributes: String,
